Reply Copy Avoidance

+++ src/defrag.c
@@ -284,7 +284,7 @@ void *activeDefragHfieldAndUpdateRef(void *ptr, void *privdata) {
  * reference count is not 1, in these cases, the caller must explicitly pass
  * in the reference count, otherwise defragmentation will not be performed.
  * Note that the caller is responsible for updating any other references to the robj. */
-robj *activeDefragStringObEx(robj* ob, int expected_refcount) {
+robj *activeDefragStringObEx(robj* ob, unsigned int expected_refcount) {
     robj *ret = NULL;
     if (ob->refcount!=expected_refcount)
         return NULL;
@@ -1132,7 +1132,7 @@ void defragPubsubScanCallback(void *privdata, const dictEntry *de, dictEntryLink
     dict *newclients, *clients = dictGetVal(de);
 
     /* Try to defrag the channel name. */
-    serverAssert(channel->refcount == (int)dictSize(clients) + 1);
+    serverAssert(channel->refcount == dictSize(clients) + 1);
     newchannel = activeDefragStringObEx(channel, dictSize(clients) + 1);
     if (newchannel) {
         kvstoreDictSetKey(pubsub_channels, ctx->kvstate.slot, (dictEntry*)de, newchannel);
